+++ src/main/java/yokwe/majuro/type/ContextSubrange.java
@@ -6,16 +6,16 @@ public final class ContextSubrange implements CheckValue {
 	private static final org.slf4j.Logger logger = org.slf4j.LoggerFactory.getLogger(java.lang.invoke.MethodHandles.lookup().lookupClass());
 
 	public final String name;
-	public final long   minValue;
-	public final long   maxValue;
+	public final int    minValue;
+	public final int    maxValue;
 	
-	public ContextSubrange(String name, long minValue, long maxValue) {
+	public ContextSubrange(String name, int minValue, int maxValue) {
 		this.name     = name;
 		this.minValue = minValue;
 		this.maxValue = maxValue;
 	}
 	
-	public void check(long value) {
+	public void check(int value) {
 		if (minValue <= value && value <= maxValue) return;
 		logger.error("Unexpected");
 		logger.error("  class     {}", name);
